- Started adding action function declarations to objects. - Added integer constant declarations to objects. - Added some new token-based functions to sc_man.cpp that know about keywords   and record proper type information, so parsers don't need to treat   everything as strings. - Added a simple symbol table to PClass.
SVN r394 (trunk)

+++ src/s_environment.cpp
@@ -491,8 +491,8 @@ static void ReadEAX ()
 			{
 				SC_MustGetFloat ();
 				props.*EAXFields[i].Float = clamp (sc_Float,
-					float(EAXFields[i].Min)/1000.f,
-					float(EAXFields[i].Max)/1000.f);
+					double(EAXFields[i].Min)/1000,
+					double(EAXFields[i].Max)/1000);
 			}
 			else if (EAXFields[i].Int)
 			{
